refactor: define next shot button as component
