+++ app/[teamId]/[recordId]/components/sheet/types.ts
@@ -1,3 +1,5 @@
+import { RecordDetail } from "@/types/model";
+
 export type Canvas2D = CanvasRenderingContext2D;
 export type Dimensions = { width: number; height: number };
 
@@ -30,7 +32,9 @@ export interface SheetProps {
   className?: string;
   interactive?: boolean;
   onStonePositionChange: (endIndex: number, shotIndex: number, isFriendStone: boolean, newPosition: Coordinate) => void;
+  record: RecordDetail;
+  setRecord: React.Dispatch<React.SetStateAction<RecordDetail>>;
   selectedEndIndex: number;
   selectedShotIndex: number;
-  createNextShot: () => void;
+  setSelectedShotIndex: React.Dispatch<React.SetStateAction<number>>;
 }
